requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.5.0
requests==2.31.0
Werkzeug==2.3.7
gunicorn==20.1.0
python-dotenv==1.0.0
setuptools>=65.0.0
